Reorganize namespace structure Bump version
