+++ silkrpc/commands/rpc_api.hpp
@@ -26,7 +26,7 @@
 #include <silkrpc/commands/tg_api.hpp>
 #include <silkrpc/commands/trace_api.hpp>
 #include <silkrpc/commands/web3_api.hpp>
-#include <silkrpc/ethdb/kv/database.hpp>
+#include <silkrpc/ethdb/database.hpp>
 
 namespace silkrpc::http { class RequestHandler; }
 
@@ -34,7 +34,7 @@ namespace silkrpc::commands {
 
 class RpcApi : protected EthereumRpcApi, NetRpcApi, Web3RpcApi, DebugRpcApi, ParityRpcApi, TurboGethRpcApi, TraceRpcApi {
 public:
-    explicit RpcApi(std::unique_ptr<ethdb::kv::Database>& database)
+    explicit RpcApi(std::unique_ptr<ethdb::Database>& database)
     : EthereumRpcApi{database}, NetRpcApi{}, Web3RpcApi{database}, DebugRpcApi{database},
         ParityRpcApi{database}, TurboGethRpcApi{database}, TraceRpcApi{database} {}
     virtual ~RpcApi() {}
