+++ silkrpc/commands/trace_api.hpp
@@ -26,7 +26,7 @@
 
 #include <silkrpc/core/rawdb/accessors.hpp>
 #include <silkrpc/json/types.hpp>
-#include <silkrpc/ethdb/kv/database.hpp>
+#include <silkrpc/ethdb/database.hpp>
 
 namespace silkrpc::http { class RequestHandler; }
 
@@ -34,7 +34,7 @@ namespace silkrpc::commands {
 
 class TraceRpcApi {
 public:
-    explicit TraceRpcApi(std::unique_ptr<ethdb::kv::Database>& database) : database_(database) {}
+    explicit TraceRpcApi(std::unique_ptr<ethdb::Database>& database) : database_(database) {}
     virtual ~TraceRpcApi() {}
 
     TraceRpcApi(const TraceRpcApi&) = delete;
@@ -52,7 +52,7 @@ protected:
     asio::awaitable<void> handle_trace_transaction(const nlohmann::json& request, nlohmann::json& reply);
 
 private:
-    std::unique_ptr<ethdb::kv::Database>& database_;
+    std::unique_ptr<ethdb::Database>& database_;
 
     friend class silkrpc::http::RequestHandler;
 };
