+++ silkrpc/commands/web3_api.hpp
@@ -26,7 +26,7 @@
 
 #include <silkrpc/core/rawdb/accessors.hpp>
 #include <silkrpc/json/types.hpp>
-#include <silkrpc/ethdb/kv/database.hpp>
+#include <silkrpc/ethdb/database.hpp>
 
 namespace silkrpc::http { class RequestHandler; }
 
@@ -34,7 +34,7 @@ namespace silkrpc::commands {
 
 class Web3RpcApi {
 public:
-    explicit Web3RpcApi(std::unique_ptr<ethdb::kv::Database>& database) : database_(database) {}
+    explicit Web3RpcApi(std::unique_ptr<ethdb::Database>& database) : database_(database) {}
     virtual ~Web3RpcApi() {}
 
     Web3RpcApi(const Web3RpcApi&) = delete;
@@ -45,7 +45,7 @@ protected:
     asio::awaitable<void> handle_web3_sha3(const nlohmann::json& request, nlohmann::json& reply);
 
 private:
-    std::unique_ptr<ethdb::kv::Database>& database_;
+    std::unique_ptr<ethdb::Database>& database_;
 
     friend class silkrpc::http::RequestHandler;
 };
