+++ silkrpc/http/request_handler.hpp
@@ -32,7 +32,7 @@
 #include <asio/awaitable.hpp>
 
 #include <silkrpc/commands/rpc_api.hpp>
-#include <silkrpc/ethdb/kv/database.hpp>
+#include <silkrpc/ethdb/database.hpp>
 
 namespace silkrpc::http {
 
@@ -44,14 +44,14 @@ public:
     RequestHandler(const RequestHandler&) = delete;
     RequestHandler& operator=(const RequestHandler&) = delete;
 
-    explicit RequestHandler(std::unique_ptr<ethdb::kv::Database>& database)
+    explicit RequestHandler(std::unique_ptr<ethdb::Database>& database)
     : database_(database), rpc_api_{database} {}
     virtual ~RequestHandler() {}
 
     asio::awaitable<void> handle_request(const Request& request, Reply& reply);
 
 private:
-    std::unique_ptr<ethdb::kv::Database>& database_;
+    std::unique_ptr<ethdb::Database>& database_;
     commands::RpcApi rpc_api_;
 
     typedef asio::awaitable<void> (commands::RpcApi::*HandleMethod)(const nlohmann::json&, nlohmann::json&);
